Add styling to show page

+++ db/schema.rb
@@ -16,8 +16,8 @@ ActiveRecord::Schema.define(version: 2021_06_19_164726) do
   enable_extension "plpgsql"
 
   create_table "article_categories", force: :cascade do |t|
-    t.integer "article_id", null: false
-    t.integer "category_id", null: false
+    t.bigint "article_id", null: false
+    t.bigint "category_id", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.index ["article_id"], name: "index_article_categories_on_article_id"
@@ -25,7 +25,7 @@ ActiveRecord::Schema.define(version: 2021_06_19_164726) do
   end
 
   create_table "articles", force: :cascade do |t|
-    t.integer "user_id", null: false
+    t.bigint "user_id", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.string "title"
@@ -43,8 +43,8 @@ ActiveRecord::Schema.define(version: 2021_06_19_164726) do
   end
 
   create_table "comments", force: :cascade do |t|
-    t.integer "user_id", null: false
-    t.integer "article_id", null: false
+    t.bigint "user_id", null: false
+    t.bigint "article_id", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.text "body"
@@ -66,8 +66,8 @@ ActiveRecord::Schema.define(version: 2021_06_19_164726) do
   end
 
   create_table "votes", force: :cascade do |t|
-    t.integer "user_id", null: false
-    t.integer "article_id", null: false
+    t.bigint "user_id", null: false
+    t.bigint "article_id", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.index ["article_id"], name: "index_votes_on_article_id"
